refactor(components): restructure Text and Link components
