+++ packages/react/src/components/text/text.hook.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import type { ReactNode, ElementType } from "react";
 
-import { createComponentProps } from "../utils";
+import { createComponentProps } from "../../utils";
 
 import { textVariants } from "./text.variants";
 import type { TextVariantProps } from "./text.variants";
