+++ packages/react/src/components/text/text.tsx
