+++ packages/react/src/index.ts
@@ -1,5 +1,1 @@
-export { default as Text } from "./text";
-export type * from "./text";
-
-export { default as Link } from "./link";
-export type * from "./link";
+export * from "./components";
