+++ turbo/generators/templates/component/component.hook.hbs
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import type { ReactNode, ElementType } from "react";
 
-import { createComponentProps } from "../utils";
+import { createComponentProps } from "../../utils";
 
 import { {{camelCase name}}Variants } from "./{{kebabCase name}}.variants";
 import type { {{pascalCase name}}VariantProps } from "./{{kebabCase name}}.variants";
